Tests/LibWeb/Text/expected/Editing/click-editable-hidden-on-focus.txt

selection: #text("edit me")@4
active element: editor
